more tidying, give cache stuff own file
